+++ arcade-ui/src/arcade-ui-react-shim.cjs
@@ -1,0 +1,6 @@
+'use strict'
+
+/**
+ * CommonJS companion for `@davide03memoli/arcade-ui/react` when consumed via require().
+ */
+module.exports = {}